# src/main/java/hu/cubixwebshop/catalogservice/openapi/api/CategoryControllerApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package hu.cubixwebshop.catalogservice.openapi.api;

import hu.cubixwebshop.catalog_service.api.ApiUtil;
import hu.cubixwebshop.catalogservice.openapi.model.CategoryDto;
import hu.cubixwebshop.catalogservice.openapi.model.HistoryDataCategoryDto;
import io.swagger.annotations.ApiParam;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;

import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-03T13:33:12.168299443+01:00[Europe/Budapest]")
@Validated
@Tag(name = "CategoryController", description = "the CategoryController API")
public interface CategoryControllerApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /openapi/categories
     *
     * @param categoryDto  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createCategory",
        tags = { "categoryController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/categories",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<CategoryDto> createCategory(
        @Parameter(name = "CategoryDto", description = "", required = true) @Valid @RequestBody CategoryDto categoryDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"categoryname\" : \"\", \"id\" : \"\", \"products\" : { \"price\" : \"\", \"productname\" : \"\", \"id\" : \"\" } }";
                    hu.cubixwebshop.catalog_service.api.ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /openapi/categories/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteCategory",
        tags = { "categoryController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/categories/{id}"
    )
    default ResponseEntity<Void> deleteCategory(@ApiParam(value = "", required = true) @PathVariable("id") Long id

    ){
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

    /**
     * GET /openapi/categories/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getCategoryById",
        tags = { "categoryController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/categories/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<CategoryDto> getCategoryById(@ApiParam(value = "", required = true) @PathVariable("id") Long id

    ){
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"categoryname\" : \"\", \"id\" : \"\", \"products\" : { \"price\" : \"\", \"productname\" : \"\", \"id\" : \"\" } }";
                    hu.cubixwebshop.catalog_service.api.ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /openapi/categories/{id}/history
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getCategoryHistoryById",
        tags = { "categoryController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/categories/{id}/history",
        produces = { "application/json" }
    )
    default ResponseEntity<List<HistoryDataCategoryDto>> getCategoryHistoryById(@ApiParam(value = "", required = true) @PathVariable("id") Long id

    )
    {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /openapi/categories/{id}
     *
     * @param id  (required)
     * @param categoryDto  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "modifyCategory",
        tags = { "categoryController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = CategoryDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/categories/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<CategoryDto> modifyCategory(@ApiParam(value = "", required = true) @PathVariable("id") Long id

            ,

                                                       @ApiParam(value = "", required = true )   @Valid @RequestBody CategoryDto categoryDto){
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"categoryname\" : \"\", \"id\" : \"\", \"products\" : { \"price\" : \"\", \"productname\" : \"\", \"id\" : \"\" } }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /api/categories/search
     *
     * @param full  (optional)
     * @param page  (optional)
     * @param size  (optional)
     * @param sort  (optional)
     * @return OK (status code 200)
     */
    @Operation(
            operationId = "search",
            tags = { "categoryController" },
            responses = {
                    @ApiResponse(responseCode = "200", description = "OK", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/api/categories/search",
            produces = { "application/json" }
    )
    default ResponseEntity<List<CategoryDto>>  search(
            @Parameter(name = "full", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "full", required = false) Boolean full,
            @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
            @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
            @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) List<String>  sort
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


}