# src/main/java/hu/cubixwebshop/catalogservice/openapi/api/ProductControllerApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package hu.cubixwebshop.catalogservice.openapi.api;

import hu.cubixwebshop.catalog_service.api.ApiUtil;
import hu.cubixwebshop.catalogservice.openapi.model.CategoryDto;
import hu.cubixwebshop.catalogservice.openapi.model.HistoryDataCategoryDto;
import hu.cubixwebshop.catalogservice.openapi.model.HistoryDataProductDto;
import hu.cubixwebshop.catalogservice.openapi.model.ProductDto;
import io.swagger.annotations.ApiParam;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;

import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-01-03T13:33:12.168299443+01:00[Europe/Budapest]")
@Validated
@Tag(name = "ProductController", description = "the ProductController API")
public interface ProductControllerApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /openapi/products
     *
     * @param productDto  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "createProduct",
        tags = { "productController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/products",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ProductDto> createProduct(@Valid ProductDto productDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"price\" : \"\", \"productname\" : \"\", \"id\" : \"\" }";
                    hu.cubixwebshop.catalog_service.api.ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * DELETE /openapi/products/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "deleteProduct",
        tags = { "productController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/api/products/{id}"
    )
    default ResponseEntity<Void> deleteProduct(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Object id
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /openapi/products
     *
     * @param full  (optional)
     * @param page  (optional)
     * @param size  (optional)
     * @param sort  (optional)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "findProductAll",
        tags = { "productController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/products",
        produces = { "application/json" }
    )
    default ResponseEntity<List<ProductDto>> findProductAll(
        @Parameter(name = "full", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "full", required = false) Boolean full,
        @Parameter(name = "page", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false) Integer page,
        @Parameter(name = "size", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false) Integer size,
        @Parameter(name = "sort", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) List<String>  sort
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /openapi/products/{id}
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getProductById",
        tags = { "productController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/products/{id}",
        produces = { "application/json" }
    )
    default ResponseEntity<ProductDto> getProductById(@ApiParam(value = "", required = true) @PathVariable("id") Long id

    ){
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"price\" : \"\", \"productname\" : \"\", \"id\" : \"\" }";
                    hu.cubixwebshop.catalog_service.api.ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /openapi/products/{id}/history
     *
     * @param id  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "getProductsHistoryById",
        tags = { "productController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/api/products/{id}/history",
        produces = { "application/json" }
    )
    default ResponseEntity<List<HistoryDataProductDto>> getProductsHistoryById(@ApiParam(value = "", required = true) @PathVariable("id") Long id

    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * POST /openapi/products/{id}/price/{price}
     *
     * @param id  (required)
     * @param price  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "modifyPriceProduce",
        tags = { "productController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = Object.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/api/products/{id}/price/{price}",
        produces = { "application/json" }
    )
    default ResponseEntity<Void> modifyPriceProduce(
        @Parameter(name = "id", description = "", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "price", description = "", required = true, in = ParameterIn.PATH) @PathVariable("price") double price
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * PUT /openapi/products/{id}
     *
     * @param id  (required)
     * @param productDto  (required)
     * @return OK (status code 200)
     */
    @Operation(
        operationId = "modifyProduct",
        tags = { "productController" },
        responses = {
            @ApiResponse(responseCode = "200", description = "OK", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductDto.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/api/products/{id}",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ProductDto> modifyProduct(@ApiParam(value = "", required = true) @PathVariable("id") Long id,
                                                     @ApiParam(value = "", required = true )   @Valid @RequestBody ProductDto productDto)
     {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"price\" : \"\", \"productname\" : \"\", \"id\" : \"\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


}